requirements.txt
streamlit
yfinance
plotly
pandas
python-dotenv